decoupled settings/signals from AudioOutput/Player/etc set down groundwork for sox pipeline set down groundwork for fast adapter selection
