+++ .gitignore
@@ -158,3 +158,7 @@ venv.bak/
 .mypy_cache/
 
 /scrap/
+/audio_player/lib/audio_file/adapters/sox/
+/audio_player/lib/audio_file/adapters/sox/sox_adapter.py
+/audio_player/lib/audio_file/adapters/sox/__init__.py
+/audio_player/lib/audio_file/adapters/sox/sox_process.py
