+++ audio_player/__init__.py
@@ -1,4 +1,24 @@
-from .audio_player import AudioPlayer
-from .audio_player_process_interface import AudioPlayerProcessInterface
-from .lib.exceptions import UnableToOpenFileError
-# from .audio_player_process import AudioPlayerProcess
+from typing import TYPE_CHECKING
+
+if TYPE_CHECKING:
+    from .audio_player import AudioPlayer
+    from .audio_player_process_interface import AudioPlayerProcessInterface
+
+__all__ = [
+    'AudioPlayer',
+    'AudioPlayerProcessInterface'
+]
+
+
+def __getattr__(name):
+    if name == 'AudioPlayer':
+        from .audio_player import AudioPlayer
+        return AudioPlayer
+
+    elif name == 'AudioPlayerProcessInterface':
+        from .audio_player_process_interface import AudioPlayerProcessInterface
+        return AudioPlayerProcessInterface
+
+    else:
+        raise ImportError
+
