+++ audio_player/audio_player_settings.py
@@ -1,0 +1,7 @@
+from .audio_output_settings import Settings as AudioOutputSettings
+
+
+class Settings(AudioOutputSettings):
+    def __init__(self):
+        super().__init__()
+        self.looping = self.add_setting('looping', False)
