+++ audio_player/audio_processing.py
@@ -2,7 +2,7 @@ import numpy as np
 from numba import jit
 
 
-@jit(nopython=True, nogil=True, parallel=True)
+@jit(nopython=True, nogil=True, parallel=True, cache=True, no_rewrites=True)
 def process_vpb(
         data: np.ndarray,
         volume: float,
@@ -13,7 +13,7 @@ def process_vpb(
         data *= volume
 
     if volume:
-        if pan and data.shape[1] == 2:
+        if pan and data.ndim == 2 and data.shape[1] == 2:
             # if data.shape[1] == 1:
             #     data = np.repeat(data[:, np.newaxis], 2, 1)
             abs_pan = abs(pan)
