+++ examples/qt_example/clickable_slider.py
@@ -1,4 +1,4 @@
-import qt
+from . import qt
 
 
 class ClickableSlider(qt.QSlider):
