+++ examples/qt_example/q_audio_player_process_interface.py
@@ -1,0 +1,37 @@
+from . import qt
+from audio_player import AudioPlayerProcessInterface
+# from audio_player import AudioPlayer
+
+
+class QAudioPlayerProcessInterface(AudioPlayerProcessInterface):
+    class QSignals(qt.QObject):
+        file_changed = qt.pyqtSignal(object)
+        state_changed = qt.pyqtSignal(str)
+        duration_changed = qt.pyqtSignal(float)
+        position_changed = qt.pyqtSignal(float)
+
+    def __init__(self):
+        super().__init__()
+        self.qsignals = QAudioPlayerProcessInterface.QSignals()
+
+        self.signals.file_changed.connect(self.qsignals.file_changed.emit)
+        self.signals.state_changed.connect(self.qsignals.state_changed.emit)
+        self.signals.position_changed.connect(self.qsignals.position_changed.emit)
+        self.signals.duration_changed.connect(self.qsignals.duration_changed.emit)
+
+
+# class QAudioPlayerProcessInterface(AudioPlayer):
+#     class QSignals(qt.QObject):
+#         file_changed = qt.pyqtSignal(object)
+#         state_changed = qt.pyqtSignal(str)
+#         duration_changed = qt.pyqtSignal(float)
+#         position_changed = qt.pyqtSignal(float)
+#
+#     def __init__(self):
+#         super().__init__()
+#         self.qsignals = QAudioPlayerProcessInterface.QSignals()
+#
+#         self.signals.file_changed.connect(self.qsignals.file_changed.emit)
+#         self.signals.state_changed.connect(self.qsignals.state_changed.emit)
+#         self.signals.position_changed.connect(self.qsignals.position_changed.emit)
+#         self.signals.duration_changed.connect(self.qsignals.duration_changed.emit)
