+++ examples/qt_example/qt.py
@@ -2,3 +2,7 @@ from PyQt5.Qt import *
 from PyQt5.Qt import Qt as Constants
 
 constants = Constants
+
+
+# def __getattr__(name):
+#     print('qt', 'getattr', name)
